Move classes related to security in ...auth package to ...security package
