+++ common/src/test/java/io/pravega/common/security/ZKTLSUtilsTest.java
@@ -7,7 +7,7 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.pravega.common.auth;
+package io.pravega.common.security;
 
 import org.junit.Assert;
 import org.junit.Test;
