+++ controller/src/main/java/io/pravega/controller/server/rest/RESTServer.java
@@ -12,7 +12,7 @@ package io.pravega.controller.server.rest;
 import com.google.common.util.concurrent.AbstractIdleService;
 import io.pravega.client.netty.impl.ConnectionFactory;
 import io.pravega.common.LoggerHelpers;
-import io.pravega.common.auth.JKSHelper;
+import io.pravega.common.security.JKSHelper;
 import io.pravega.controller.server.ControllerService;
 import io.pravega.controller.server.eventProcessor.LocalController;
 import io.pravega.controller.server.rest.resources.PingImpl;
