+++ controller/src/main/java/io/pravega/controller/store/client/StoreClientFactory.java
@@ -12,8 +12,8 @@ package io.pravega.controller.store.client;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import io.pravega.common.Exceptions;
-import io.pravega.common.auth.JKSHelper;
-import io.pravega.common.auth.ZKTLSUtils;
+import io.pravega.common.security.JKSHelper;
+import io.pravega.common.security.ZKTLSUtils;
 import lombok.AccessLevel;
 import lombok.Setter;
 import lombok.Synchronized;
