+++ standalone/src/main/java/io/pravega/local/InProcPravegaCluster.java
@@ -12,7 +12,7 @@ package io.pravega.local;
 import com.google.common.base.Preconditions;
 import io.pravega.client.stream.impl.Credentials;
 import io.pravega.client.stream.impl.DefaultCredentials;
-import io.pravega.common.auth.ZKTLSUtils;
+import io.pravega.common.security.ZKTLSUtils;
 import com.google.common.base.Strings;
 import io.pravega.controller.server.ControllerServiceConfig;
 import io.pravega.controller.server.ControllerServiceMain;
